Moved part of the code from CadastroAtletaForm to CadastroAtletas.
